max_message_size is not available in released versions of Mailman
Since max_message_size is not available in the released versions of
Mailman (<=3.1.1), the max_message_size field creates errors when set to a
non-none value. However, since it was marked as a required field, not setting it
would raise errors too!

This is a temporary fix that sets it to required=False so that the rest of the
fields can be updated by setting max_message_size to None.