Metadata and docs changes for Python 3 support.
Git head test should not use Django's latest version.

Fix Postorius's tests.

This fixes some of the failing git tests due to some of the change in repr of
the objects. This commits changes the comparison to List's fqdn_listname instead
of repr, which isn't even supposed to machine readable string.