require_GET doesn't allow HEAD requests which WSGI Servers depend on
require_GET only allows GET requests, but require_safe allows GET and HEAD. HEAD
is sometimes required by WSGI servers to properly generate the response. With
require_GET, the WSGI server doesn't return the response properly, but, Django's
test server allows that happily.

It is hard to reproduce this in a Django test case.

In case someone is looking for the differences in between HEAD and GET requests,
this blog post is quite informative:
http://blog.dscpl.com.au/2009/10/wsgi-issues-with-http-head-requests.html